+++ src/features/Calendar/hooks/index.ts
@@ -9,10 +9,10 @@ export { useCalendarEvents } from './useCalendarEvents'
 export { useCalendarKeyDelete } from './useCalendarKeyDelete'
 export { useCalendarModal } from './useCalendarModal'
 export { useCalendarNavigation } from './useCalendarNavigation'
-export { useCalendarPortals } from './useCalendarPortals'
 export { useCalendarRbcProps } from './useCalendarRbcProps'
 export { useCalendarResponsive } from './useCalendarResponsive'
 export { useCalendarSelection } from './useCalendarSelection'
 export { useCalendarTodoTimingPatch } from './useCalendarTodoTimingPatch'
+export { useCustomCalendarController } from './useCustomCalendarController'
 export { useDayViewHandlers } from './useDayViewHandlers'
 export { useStoredCalendarView } from './useStoredCalendarView'
